+++ app/Http/Controllers/Api/V1/QuestionController.php
@@ -16,6 +16,8 @@ class QuestionController extends Controller
 {
     public function index(Request $request): AnonymousResourceCollection
     {
+        $this->authorize('viewAny', Question::class);
+
         $itemsPerPage = $request->input('itemsPerPage', self::ITEMS_PER_PAGE);
 
         $builder = QueryBuilder::for(Question::class)
@@ -29,6 +31,8 @@ class QuestionController extends Controller
 
     public function store(QuestionRequest $request): JsonResponse
     {
+        $this->authorize('create', Question::class);
+
         $question = Question::create($request->validated());
 
         return (new QuestionResource($question))
@@ -38,11 +42,15 @@ class QuestionController extends Controller
 
     public function show(Question $question): QuestionResource
     {
+        $this->authorize('view', Question::class);
+
         return new QuestionResource($question);
     }
 
     public function update(QuestionRequest $request, Question $question): JsonResponse
     {
+        $this->authorize('update', Question::class);
+
         $question->update($request->validated());
 
         return (new QuestionResource($question))
@@ -52,8 +60,26 @@ class QuestionController extends Controller
 
     public function destroy(Question $question): JsonResponse
     {
+        $this->authorize('delete', Question::class);
+
         $question->delete();
 
         return response()->json([], Response::HTTP_NO_CONTENT);
+    }
+
+    //todo check for best practice
+    public function answers(Question $question)
+    {
+        $this->authorize('viewAnswers', Question::class);
+
+        $answers = $question->answers->map(function ($answer) {
+            return [
+                'id' => $answer->id,
+                'text' => $answer->text,
+                'question_id' => $answer->question_id,
+            ];
+        });
+
+        return $answers;
     }
 }
